requirements.txt
keyboard==0.13.5
PyDirectInput==1.0.4
PyYAML==6.0.1
